% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaNetis.R
\name{runMetaNetis}
\alias{runMetaNetis}
\title{Launch Shiny App for MetaNetis}
\usage{
runMetaNetis()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for MetaNetis.
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

MetaNetis::runMetaNetis()

}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
Yunze Du, \email{yunze.du@mail.utoronto.ca}
}
