% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetAltBaseline.R
\name{SetAltBaseline}
\alias{SetAltBaseline}
\title{Set an Alternative Reference Range Baseline}
\usage{
SetAltBaseline(data_source)
}
\arguments{
\item{data_source}{Either a character string specifying the file path to a CSV file
or a data frame object containing the alternative reference range data.}
}
\description{
Loads a user-provided data source (either a file path to a CSV or a data frame object)
and assigns it to the name \code{reference_ranges_df} in the global environment,
overriding the default packaged data.

This function is useful for researchers who wish to use institution-specific,
population-specific, or more up-to-date reference ranges than those bundled
with the package.
}
\details{
The replacement data frame MUST have the following column names (matching those
generated by the data cleaning pipeline):
\itemize{
  \item \code{HMDB_ID}: Unique identifier of the metabolite.
  \item \code{Metabolite_Name}: Common name of the metabolite.
  \item \code{Sample_Type}: Standardized biological fluid type (e.g., Blood/Serum/Plasma, CSF, Urine).
  \item \code{Subject_Age_Description}: The original age group description from the HMDB dataset.
  \item \code{Min_Age(year)}: Minimum age (in years) for the applicable age group.
  \item \code{Max_Age(year)}: Maximum age (in years) for the applicable age group (\code{Inf} for general adult groups).
  \item \code{Min_Concentration(Healthy)}: Minimum concentration observed in healthy individuals within the age group.
  \item \code{Mean_Concentration(Healthy)}: Aggregated average concentration in healthy individuals within the age group.
  \item \code{Max_Concentration(Healthy)}: Maximum concentration observed in healthy individuals within the age group.
  \item \code{Unit}: The unit of measurement for concentration (e.g., uM, umol/mmol creatinine).
}
}
\examples{
# Example 1:
ref_df <- data.frame(
  HMDB_ID = c("HMDB0000190", "HMDB0000064", "HMDB0000001", "HMDB0001918"),
  Metabolite_Name = c("L-Tryptophan", "Creatinine", "1-Methylhistidine", "Glucose"),
  Sample_Type = c("Blood/Serum/Plasma", "Urine", "CSF", "Blood/Serum/Plasma"),
  Subject_Age_Description = c("Adult (>18 years old)", "Newborn (0-30 days old)", "Adolescent (13-18 years old)", "Children (1-13 years old)"),

  'Min_Age (year)' = c(18.0, 0.0, 13.0, 1.0),
  'Max_Age (year)' = c(Inf, 0.082, 18.0, 13.0), # 30 days is ~0.082 years

  'Min Concentration (Healthy)' = c(50.0, 0.15, 0.012, 3500.0),
  'Mean (Healthy)' = c(75.5, 0.25, 0.021, 5200.0),
  'Max Concentration (Healthy)' = c(100.0, 0.50, 0.030, 8000.0),
  'Concentration Unit' = c("uM", "umol/mmol creatinine", "uM", "uM"),
  stringsAsFactors = FALSE
)

SetAltBaseline(ref_df)
GetRefRanges()

}
